Refactor JsonTypeFactory
Converted interface to abstract class to better reflect 'what it is' as
opposed to 'what it does', and allow each implementation to store the
schema which it was instantiated to resolve. Improved use of prototype
scope on related Spring beans.

The app version and JAR have not been updated as no core functionality
was affected by this commit.