Populate arrays with appropriate values
Helper classes have been created which supply appropriate definitions to
JsonArrayFactory for permitted array contents. Each dialect which added
or altered array-specific keywords has a separate class; note that some
of those are distinguished by keywords which are not yet inspected by
the application.

Any implementation of Iterable<JsonSchema> can be used in place of those
classes to supply item definitions, when using the application code as a
library.